packages/python/api-client-generator/pyproject.toml
[project]
name = "openverse-api-client-generator"
version = "0.0.0"
description = "Default template for PDM package"
authors = [
    {name = "The Openverse Contributors", email = "openverse@wordpress.org"},
]
dependencies = [
    "pyyaml>=6.0.1",
    "jinja2>=3.1.4",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
distribution = false

[tool.pdm.scripts]
generate = { call = "openverse_api_client_generator.__main__:cli" }

[tool.pdm.dev-dependencies]
dev = [
    "mypy>=1.10.0",
]
